refactor: WithdrawTopAppBar → NearCancelTopAppBar로 이름 변경 및 이동
- 공통 컴포넌트로 활용하기 위해 ui/componenets로 이동
